% Generated by roxygen2 (4.0.2): do not edit by hand
\name{genernk}
\alias{genernk}
\title{ranking the genes via DNMF}
\usage{
genernk(object)
}
\arguments{
\item{object,}{a returned object of function \code{DNMF}}
}
\description{
ranking the genes via DNMF from a returned object of function \code{DNMF}.
}
\examples{
\dontrun{
genernk(dnmf_result)
}
}

