% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rnk.R
\name{rnk}
\alias{rnk}
\title{write rnk to a file from matrix W.}
\usage{
rnk(object, fn = "./tmp.rnk", type = "o2m")
}
\arguments{
\item{object}{a returned object of function \code{DNMF}}

\item{fn}{the output filename. Default is "./tmp.rnk"}

\item{type}{type o2m (Default) or o2o. to compare with multi sample labels. 
o2m means one Vs others, while o2o means one Vs another one.}
}
\description{
write a rnk file from matrix W in a returned object of function \code{DNMF}.
The rnk format is referred \href{http://www.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats}{RNK}
}
\examples{
\dontrun{
rnk(dnmf_result, fn="tmp.rnk")
}
}
