% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rnk.R
\name{rnk}
\alias{rnk}
\title{write a rnk file from matrix W.}
\usage{
rnk(object, fn = "./tmp.rnk")
}
\arguments{
\item{object,}{a returned object of function \code{DNMF}}

\item{fn,}{the output filename}
}
\description{
write a rnk file from matrix W in a returned object of function \code{DNMF}.
The rnk format is referred \href{http://www.broadinstitute.org/cancer/software/gsea/wiki/index.php/Data_formats#RNK:_Ranked_list_file_format_.28.2A.rnk.29}{RNK}
}
\examples{
\dontrun{
rnk(dnmf_result, fn="./tmp.rnk")
}
}

